Del opp innsending i et synkront og et asynkront steg (#582)
Bakgrunn
Oppslag hos PDL og BRREG for å hente person- og virksomhetsnavn i forbindelse med innsending tar lang tid, feiler sporadisk eller kan være helt nede. Dette gjør at brukerne må vente lenge, prøve flere ganger eller kanskje ikke får til i det hele tatt å sende inn inntektsmelding via nav.no

Løsning
Splitt dagens innsending i to steg:

Ta imot og lagre skjemaet brukeren sender inn.
Gi OK-respons til brukeren straks skjemaet er lagret i den nye databasekolonnen.
Dette steget forblir hetende InnsendingService, men er en nedskrellet utgave av dagens InnsendingService.
Berik og lagre inntektsmeldingen
Berik inntektsmeldingen med data fra forespørselen, eventuelle tidligere og eksterne inntektsmeldinger, PDL og BRREG.
Lagre inntektsmeldingen og send den videre (som tidligere) med INNTEKTSMELDING_MOTTATT-eventet.
Dette steget lages som en stateless-service for å gjøre oss uavhengige av Redis-timeouts, og får navnet BerikInntektsmeldingService.
Fordi vi ikke lenger kan belage oss på at brukeren sender inn på nytt dersom berikelsen feiler (slik vi kan i dagens InnsendingService), så lar vi FeilLytter plukke opp meldingen som feilet slik at berikelsen kan fortsette på et senere tidspunkt.

+++ db/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/db/tabell/InntektsmeldingEntitet.kt
@@ -10,7 +10,7 @@ import org.jetbrains.exposed.sql.json.jsonb
 
 object InntektsmeldingEntitet : Table("inntektsmelding") {
     val id =
-        integer("id").autoIncrement(
+        long("id").autoIncrement(
             idSeqName = "inntektsmelding_id_seq",
         )
     val forespoerselId = varchar(name = "forespoersel_id", length = 40) references ForespoerselEntitet.forespoerselId
